Weeks should start on Monday
With the cal command, Sunday is the first day of the week, which is from a
European point of view a bit odd. As invoking the cal command doesn't allow you
to use Monday as the first day, I added an alias for cal, which invokes now ncal
with the b option. ncal is localized by default, and the b option forces it to
use the cal month layout.