Test if gpg-agent is running by trying to kill it
The environment file of gpg-agent isn't deleted gpg-agent has stopped.
Testing if the file exists, like the site of gpg says, isn't really
helpful. I used to delete the file in /etc/rc.local, but this is
certainly a far superior solution, as it's easier and portable. On top
of that, this will restart gpg-agent in the hypothetical case it would
have crashed or been killed.