Improve way to check if gpg-agent is running
By default, the root Xsession file seems to start gpg-agent. One could
think it's a good idea to simply use this daemon. The problem is that it
only works from withing the GUI. If you use a tty, you won't be asked
for a password on the tty, but in X. It checks whether a daemon has
already been started, but this fails as they don't export the variables,
so you can't block it from starting, but removing it from
/etc/X11/Xsession.d/

Xfce starts gpg-agent as well, but only if there is no environment
variables set, so, we have no problem here.