Use gpg-agent environment file as used by Xsession
It seems like the ssh-agent of gpg-agent does show the pinentry at the
tty/display where the daemon has been started anyway.

It's a pity in fact, but there is no real universal alternative, to my
knowledge, as ssh-agent requires you to add keys each session. You can
automate this (and/or use keychain), but then you have to give the
password of all keys at login time, which is very inconvenient, as you
wont use most of them during the session.