Use default location of environment file
Use the default location of the environment file after all. It seems you
can change the display where the pinentry is shown with the command:
gpg-connect-agent updatestartuptty /bye

Using a wrapper around the ssh command makes it be changed to the
display from wherever the command is executed, exactly what I want. But,
this is impossible using the gpg-agent started by Xsession, as this is
bound to the my Xsession.