Start gvfsd before xfce4
You have to start gvfsd manually when you want it to use gpg-agent,
otherwise it seems it doesn't get the right environment variables if
started by Thunar. But, starting it with session, you sometimes get an
error, while repeatedly clicking will show you a pinentry. Starting it
with the xsession does seem to be more reliable.