Use Vundle for nearly all plugins
I think it's easier and more portable to have one plugin manager, and
use it for all plugins. Now, Vundle is used for all plugins, except:
 - YouCompleteMe
 - bicyclerepair
 - javaimports
 - bbcode

YouCompleteMe has to be compiled, and for that reason, I'll continue
using the Debian package for now. Bicyclerepair is a plugin delivered
with the tool itself, and is of no use without it. It would probably be
redundant if I would use python-mode. Javaimports is a script from the
wiki, but I suppose I could use a gist. bbcode is a slightly modified
version from a syntax file I found somewhere. I guess, this could be
done with a gist as well.