Switch to Alt-x as prefix
Ctrl-X is used for multiple functions in Vim, most of them I don't use
immediately (auto completion plugins), but sometimes I want to be able
to test it by hand. As alt isn't really used in Vim as such, this seems
to be an acceptable and quite easy to get used to replacement, I hope.