Output control
The trigger for output should be low, to be certain tmux is triggered.
The default interval (time where the output "freezes") is quite long,
and is very noticeable, and in my opinion, annoying. That's why I
reduced it to 10 ms. This gives the same responsiveness and feel as a
plain terminal.