Add key mappings for GVim
In the terminal, using Alt combined with any key, gives an escape
character and the other pressed key(s). As such, you can directly use
some normal mode keys in insert mode. I used this a lot for A and O, and
sometimes o. This doesn't work in GVim, and as I type mails in GVim,
it's handy to have these shortcuts there as well.