Stop automatic use of proxy
The benefit does seem to be to low compared to the additional time
required to recover from a connection loss, for example during a train
ride. On top of that, many sites/most sites are HTTPS nowadays, and
encrypted data doesn't compress well.

Besides, I've seen the sometimes "high" data usage is caused nearly
exclusively by visiting GitHub. It seems to have a very bad efficiency.
It might be of interest to check whether it's feasibly to rewrite url's
to the mobile website when on 3g, although that might not be very
convenient either.