Don't start gvfsd
I don't use it anymore, and might even completely remove it. I even
realized recently the only parts of Xfce I'm still using is the panel
and the power manager, and the last one only to define the default
behaviour of certain actions, and as a battery indicator. I won't
immediately switch to xmobar though, as I'll probably spend quite some
time configuring it, and I shouldn't spend it on this.