Better remapping of keys
Although some of these remapping aren't necessary on my laptop, they do
come in handy when I would want to use a normal keyboard. One could be
excluded for the normal keyboard, but for now, I don't see any harm in
reassigning the left super key to esc on a normal keyboard, it won't be
used anyway.