Always add sharp 1920x1080 settings
Some weeks ago, I thought I'd lost my HDMI to VGA converter, and ordered
a new one. Today I tested him, and apart from the fact he seems to be a
little bit sharper than the one I have. But, it was recognized as being
1920x1080 capable. However, the settings were different from the mode I
add, and the black was way too bright. I also realized that the output
of my desktop, which I connect now using the VGA input of my monitor was
way less sharp. Using these settings, this is solved.