Redefine zi to set foldmethod=syntax
Having folding sometimes comes in very handy, but in some files it can
be slow, and in most files I tend to disable it immediately. Disabling
it by default seems to be the right thing to do. However, setting the
foldmethod, it will calculate the folds, no matter whether we use them.
This kind fixes this.