Workaround for fullscreen problems Iceweasel
Recently I upgraded to Iceweasel 42, as I wanted to upgrade to a more
recent VimFx version, but since my browser freezed when entering full
screen. First I thought this to be a bug in Unipept, as I thought I had
seen an error in the console, but now it seems not.

I knew there was an issue with Iceweasel as I had the same problem when
I wanted to watch a video on a website. I thought it to be an issue with
firefox, but didn't find anything at first, until I thought of narrowing
it down to xmonad, and indeed, I found this issue:
https://bugzilla.mozilla.org/show_bug.cgi?id=1189622

This gave me the workaround I needed.