Open files in one gvim instance
When opening a file from iceweasel, I had two issues, one was the fact
it used emacs by default, which can be fixed by setting gvim as default,
but then each file is opened in it's own instance. This customized
desktop file fixes this.

As neovim doesn't support remoteservers, this isn't trivial to do with
neovim. It should be possible in combination with tmux, but I couldn't
get it to work immediately, and this will do great.