Fix listing with ag
I have found the problem with ag. When no or a relative path is given,
it works as expected, but when an absolute path is given, it depends on
wetter it's finished with a / or not. When it is, gitignore works
partially, otherwise not at all. It works when the paths have only a
depth of one. (".config" works, ".config/Qtrac Ltd." doesn't)