Reduce refresh rate of volume indicator in xmobar
I tried it now for a couple of days, and while definitely not as
convenient when changing volume, it greatly reduces the load of xmobar.
On top of that, in the end you look at it twice a day ...

So, I might consider even removing it completely. I would be able to use
the Debian package then, and by using notifications, I could have the
feedback I used to have. The only problem remaining is to see the
current volume. I should think about it.