qutebrowser: only use hphosts's ad and tracker list
The complete list includes a "warez and piracy" list. However, it
doesn't even seem to be limited as that, as checking the list, it seems
an (old?) website from the "Jong VLD" in Kortrijk is blocked, which is a
political party. Plus, I don't want piracy blocked, if I don't want
piracy, I don't visit it ... This is censorship! The ads and tracker
list seems ligit, however.

The reason it's a local file is because the adblock-update command is
blocked by the server, and I couldn't find a mirror. I have a cron
script to update this file daily.