Use readline in bsh
Recently bsh was dropped as a dependency from my installed packages, and
marked for removal, around the same time a security issue was fixed. It
has drawn by attention, and I tried it, and I think it might be handy
from time to time, to quickly test something when developing.

The default command does seem to have a huge bug: when typing something
and deleting it afterwards, the initial prompt gets deleted, and all
editing you do will be completely ... messed up. But this is quickly
fixed by just having it to use proper readline.