qutebrowser: fork process
It's annoying that if I open a link, this blocks the application if no
qutebrowser tab was open yet. Note the exit at the end, which is needed
to prevent the shell to continue running. In the case of a new tab exec
could be used, but this would require some extra dupplication of code,
and would slowdown opening a new tab.