neovim: use expandtab by default
On some places it annoys me to have tabs by default. I could set it at
those places, but I'm kind of trying to reduce the use of tabs anyway,
as spaces are more robust. And the only real benefit of tabs (custom
width) does require some effort to combine with alignment properly.