muttrc: show browser when adding attachment
Always open browser when adding an attachment to a mail. Start at the
home directory, and sort all entries alphabetically.

This means we need each browser to set it's own order. For this sake the
`o` macro has been updated. Note that when selecting `unordered`, the
ordering isn't updated, and we need to reopen the browser.