qutebrowser: implement basic usermode
Vimperator has got an usermode in which stylesheets are ignored. This
can improve readability, but also make it possible to read the page
without javascript.

A basic regex to remove the stylesheets is easily written, but we also
have the issue of javascript. I do run javascript on local files, but
that's not exactly wanted when reading the page in usermode ...