qutebrowser: save session
qutebrowser, or better qtwebkit, crashes often when resuming my laptop.
It in general handles loss of connection badly. Sometimes, this happens
while quite a lot of tabs are opened, which are lost. By saving the
session, and deleting it when quiting normally, I kind of get the
behaviour of Firefox.