xmobar: fetch temperature from KMI
The USA government site xmobar uses for it's weather plugin doesn't
provide data from any of the weather stations in and around Ghent. I
used the one of Ostend, as it also provided a middle ground between
Ghent and Veurne. However, the difference today got quite big and
annoyed me.

So, I searched for an alternative weather station, bus as said, there
isn't one supported by the service used by xmobar. A quick hack is just
fetching the temperature from Melle on the website of the KMI.

It might be interesting to check if the same is possible with
buienradar, as that's what I already use for `weather-hourly`.