xmonad: replace gmrun with native prompt
I don't use it a lot, but sometimes it's handy to execute a random
command. That's possible with rifle as well, but it may not always be
useful to include it for later use.

This does everything I used of gmrun, without the need of an additional
dependency. Well, everything except searching in history.