zplug: update installation of zplug
For some reason, I never changed the installation procedure of zplug. It
still initially fetched 2.0, and the installed/updated to the latest
version. Well, currently, it isn't updated because of
https://github.com/zplug/zplug/issues/244.

Anyway, I had an issue, which wasn't due to zplug in the end, because of
which I did some testing. I did a measurement, and loading zsh seems to
have slowed down quite a lot since v2. I should confirm if this is true,
and report it.