xclipman: simple script to replace clipit
`clipit --daemon` checks every 500 ms if there is something on the
clipboard, of there isn't it restores the last known state. This,
however, is very inefficient and really uses a lot of battery.

We could just copy it using `xsel` or `xclip`, which run in the
background until they are finished. This could be put in the loop if we
prevent `xsel` or `xclip` for forking into the background.

`xsel` has an argument to do just that, `-n`, but it doesn't do anything
when the input it's given is empty. Thus, I would need to check whether
the clipboard is empty, and if it is put something in it. That's quite a
hack.

`xclip` does have the `-quiet` argument which does run in the
foreground. It also shows some informational messages, but apart from
that, it does what I want it to do.