LICENSE: MIT
One of my longest running items on my TODO list: adding a LICENSE (and
README) to my dotfiles. After all, without a license, nobody else is,
legally, allowed to use (some of) it. If it weren't for the Terms of
Service of GitHub, it's technically illegal to look ...

Anyway, copyleft doesn't make sense for dotfiles. Well, I mean, in a
sense it does make sense. I do like when others share their dotfiles
with me, but it isn't word the trouble in case of dotfiles. Anyway, most
dotfiles aren't copyleft, and it kind of defeats the purpose of pushing
them to a public repo if they can't be used.

An other potential issue would be patents, which is why the FSF
recommends to use the Apache license instead. But I don't think that's
an issue for my dotfiles, and anyway, they (shouldn't be) legal in
Belgium/Europe.

MIT seems quite dominant when it comes to dotfiles, so, MIT it is.