zsh: use zsh completion for youtube-dl
The Debian package of youtube-dl now includes the zsh autocompletion
files, and thus there is no need to use the bash completion of the
locally installed version of youtube-dl.

On the other hand, this requires the Debian package to be present, while
I use the pip version. Maybe I should consider to revert this change,
and de-install the Debian package?