xsession: use gtk2 in qt
The `qt5-style-plugins` package has been updated, and now contains the
gtk2 style and platformtheme.

There are a couple of differences between the Clearlooks and Qt's
Cleanlooks. To have a more uniform theming across applications, I
decided to go back and use the gtk2 style. The gtk2 platformtheme is
required by the gtk2 style.