getmails: don't sync work mails on private laptop
notmuch doesn't sync the tags to the emails, and is thus not synced.
Syncing the tags could be done using e.g. git, but it is quite complex,
will generate quite a lot of overhead, and there is quite a risk on
conflicts and corruption.

And then I thought: is it worth it? An urgent mail outside of working
hours should probably be sent to my private mail anyway. If I'm
expecting a mail, I can always check using the webmail client.

If notmuch would support syncing the tags to the mail, it might be
interesting though. I'm not sure, but I seem to recall that sup can do
something like that using a custom header. It's more difficult for
notmuch though, as the client will need to remove this information when
sending a reply.