xsession: disable keyboard and start kodi on eeepc
Since a couple of weeks, my eeepc is dying. Well, that is, the keyboard
is broken. Sometimes some/all keys don't work, or there are random
key presses. So, I need to disable the keyboard. If so, anything but
kodi is impossible. On top of that, since I moved to my studio in July,
I haven't done anything else either, I just use my laptop.