neovim: use ctrlp-py-matcher
With the default matcher, I often have that I typed the whole filename,
but still I haven't got the file, because it matches a letter here and
there in the path of other files.

I also tried cpsm, which should be faster, but typing e.g. `.vimrc`
didn't give me my vimrc file. It did with `.vimrc` though, but I usually
do not type dot of my dotfiles.

The only drawback of ctrlp-py-matcher seems that it highlights the whole
matching area, not just the matching characters like the default ctrlp
matcher.