xclipman: ignore non-string contents
Check if getting the contents of the clipboard was successful. This will
fail if the clipboard is empty (which will usually only happen at login)
and when the content isn't a string.

In that case, we should do nothing, and just wait till we have something
useful in the clipboard again. If we ignore this, we will wipe out the
clipboard, which makes it impossible to copy non-string objects in any
program.

Note that the xclip pipe in front of the loop prevents us getting in a
busy loop at login when the clipboard is empty.