neovim: fix version of vim-airline
I'm fed up with vim-airline breaking once a month or so. Actually, it's
the second time this week!

vim-plug allows me to track releases (`'tag': '*'`), but they
(vim-airline) rarely do a new release, and there have been a couple of
important changes since (async, neomake, ...). It works now though, and
if there is an important change of fix, I can update the commit.