neovim: disable some special glyphs in airline
Sadly there is no real separator for warnings in vim-airline. Spaces are
doubled, because the space symbol is added, while it seems to be trimmed
when there is a neomake warning. This seems to be the best thing then.
It only takes ONE space, and it's clean.

I also disabled the maxlinenr and notexists symbols. The notexists
symbol kind of annoys me because it also shows up when editing the
commit message (!), and the maxlinenr is kind of useless. There already
is a separator, and I think it's clear enough what's the linenr without
it.