urxvt: tryout Ubuntu Mono
Seems to take less vertical space, and more readable at small sizes
compared to DejaVu Sans Mono. Note that I'm using the Powerline glyphs
of DejaVu Sans Mono, as the glyphs of the patched font for Ubuntu Mono
are not sized correctly.

DejaVu Sans Mono is the fallback for many Unicode glyphs, both for
DejaVu Sans Mono and Ubuntu Mono, but as it's not monospace, some glyphs
can't be shown. By adding it explicitly at a smaller font size, those
glyphs are now supported.