dmenu: make rofi behave synchronously
It has become muscle memory to type `<mod-m><return><return>` to return
to my laptop screen, but when rofi behaves asynchronously, it actually
seems slower for small menus, and I have to wait between my presses.
Possibly it's not really slower, but just doesn't handle the selection
of an item before the elements are loaded.