mutt: mark github notifications as read
The GitHub notification mails (if I receive them ...) have a hidden
image to automatically mark the notification as read on GitHub.

The ghn-script just decodes the message, takes the HTML part and
searches for the URL, which is then opened in curl. To minimize the time
to open the message in mutt, after decoding the script continues in
background. Executing the whole script in background isn't possible, we
lose the input.