neovim: try no to close agenda
dotoo agenda is opened in a preview window. This test isn't perfect,
but there is no efficient way to get the window number of a preview
window, which is needed to efficiently check a local variable.

Alternatively, it might be interesting to test using `getbufvar` and
`bufname`, testing for a tmp file.