weechat: get notifications for all messages
This script makes it possible to receive notifications for all messages
of certain channels.

There is a script for this at
https://weechat.org/scripts/source/chan_hl.tcl.html which works, but it
prints annoying notifications in whatever weechat buffer you're in, and
it only works for irc.

This could also be done by by defining a highlight regex, but then every
message will actually be highlighted, while this only throws a signal,
catched by launcher.

Note that this script is the reason I replaced lnotify. I could have
extended lnotify, but in the end, it's probably better to just use the