qutebrowser: make ecosia default browser
I prefer the URL-based configuration of startpage, but the theme
sometimes changes to the black version. Explicitly configuring the theme
might help, but I quite like the look-and-feel of ecosia. On top of
that, their ecological goals are quite nice. The search results seem
quite on-par, although they use Bing instead of Google. So, lets give
this a try.