pulseaudio: don't switch to anything that connects
`module-switch-on-connect` is too aggressive, causing issues when the
screensaver kicks in when using HDMI. When you just hit a key or move
the cursor, the screen returns, and output is switched to HDMI. Even
when you were e.g. using headphones.

An issues unrelated to `module-switch-on-connect`, but related to the
above, is that when you plug-out the headphones, the default sink
doesn't switch to HDMI, but to the internal speakers. Those have higher
priority by default. This can be fixed using:
```sh
sudo sed -i 's/priority = 100/priority = 50/' \
/usr/share/pulseaudio/alsa-mixer/paths/analog-output-speaker.conf
```

However, this breaks bluetooth headphones. This would be perfect:
```
load-module module-switch-on-connect blacklist="^[^b][^l][^u][^e][^z]"
```

However, it doesn't seem to work. When I use it, audio still switches to
HDMI on reconnection when headphones are plugged in.