alot: drop old notmuch hack
I didn't encounter this issue for some time where one of the messages in
a thread didn't have a Date header. I encountered this sometimes with
helpscout mails, but that's no longer relevant, so I won't port this to
notmuch2 (for now). This only helped within a thread anyway, on the
thread level this really should be handled by notmuch if anything.