alot: workaround for notmuch2 segfault
After changing a couple of tags, changing the results of the current
search buffer, I get a segfault. (Issue
https://github.com/pazz/alot/issues/1504). I had a look, and there
doesn't seem to be a double free on the Python end, but I have yet to
try if I can reproduce this in C by opening a database, starting a
search, load a messages, open the database a second time, change
the tag of that message, write and close the second database, destroy
the message fetched from the first, and repeat this a couple of times.

Anyway, I still can prevent this segfault by letting the query free the
memory of it's threads, instead of destroying the thread once the Python
object gets destructed.