sway: use wpctl and limit max volume
I wanted to make the volume a round number, and was to lazy to just set
it on the command line, and thought I achieved it, turned out the
percentage is essentially just a 8-bit number ... That was not a present
experience, and not really something I want my subject my ears or
headphones to. wpctl allows to set a limit when increasing the volume,
so this is perfect.